requirements.txt
streamlit
streamlit-audiorecorder
openai>=1.2.0
